refactor(test): merge voice-chat create-session into route.test.ts (#9396)
Consolidate create-session.test.ts (14 tests) into route.test.ts (8 tests)
to produce a single test file for the voice-chat __tests__ directory,
following the epic #9391 pattern for migrating service-level tests to
route integration tests.
